minor: move trade into market
