+++ src/message/mod.rs
@@ -1,5 +1,5 @@
 use crate::market::Order;
-use crate::types::{OrderEventType, SimpleResult, Trade};
+use crate::types::{OrderEventType, SimpleResult};
 use core::cell::RefCell;
 
 use anyhow::{anyhow, Result};
@@ -53,6 +53,9 @@ pub struct OrderMessage {
     pub base: String,
     pub quote: String,
 }
+
+//re-export from market, act as TradeMessage
+pub use crate::market::Trade;
 
 #[derive(Serialize, Deserialize)]
 pub struct MessageSenderStatus {
